refactor the logging messges, supress AWS logging, and dont include local copies of javascript and stylesheets

+++ lib/static_website/compiler/compile_directory.rb
@@ -4,7 +4,7 @@ module StaticWebsite
       attr_reader :path
 
       def initialize(path, directory=true)
-        LOGGERS.each{|logger| logger.debug("Initializing StaticWebsite::Compiler::CompileDirectory with path:\n #{path}, directory: #{directory}")}
+        LOGGERS.each{|logger| logger.info("Initializing StaticWebsite::Compiler::CompileDirectory with path: #{path}, directory: #{directory}")}
         @path = directory ? path : directory_path(path)
       end
 
@@ -13,13 +13,17 @@ module StaticWebsite
       end
 
       def compile
-        LOGGERS.each{|logger| logger.debug("Making directory: #{@path}")}
-        FileUtils.mkdir_p(@path) if @path && !Dir.exists?(@path)
+        if @path && !Dir.exists?(@path)
+          LOGGERS.each { |logger| logger.info("#{self.class.to_s}##{__method__.to_s} Making directory: #{@path}") }
+          FileUtils.mkdir_p(@path)
+        end
       end
 
       def clean_up
-        LOGGERS.each{|logger| logger.debug("Removing directory: #{@path}")}
-        FileUtils.rm_rf(@path) if @path && Dir.exists?(@path)
+        if @path && Dir.exists?(@path)
+          LOGGERS.each{|logger| logger.debug("Removing directory: #{@path}")}
+          FileUtils.rm_rf(@path) if @path && Dir.exists?(@path)
+        end
       end
     end
   end
